fix(service-portal): signature collection updated ui
