+++ libs/application/templates/signature-collection/signature-list-creation/src/forms/Done.ts
@@ -63,7 +63,7 @@ export const Done: Form = buildForm({
             buildMessageWithLinkButtonField({
               id: 'done.goToServicePortal',
               title: '',
-              url: 'minarsidur/min-gogn/listar/medmaelalistar',
+              url: '/minarsidur/min-gogn/listar/medmaelalistar',
               buttonTitle: m.linkFieldButtonTitle,
               message: m.linkFieldMessage,
             }),
